update kits interface

+++ interfaces/kits/js/media/include/media_data_source_napi.h
@@ -28,9 +28,9 @@ public:
     DISALLOW_COPY_AND_MOVE(MediaDataSourceNapi);
     static napi_value Init(napi_env env, napi_value exports);
     std::shared_ptr<AVSharedMemory> GetMem();
-    int32_t ReadAt(uint32_t pos, uint32_t length);
+    int32_t ReadAt(int64_t pos, uint32_t length);
     int32_t ReadAt(uint32_t length);
-    int32_t GetSize(int32_t &size) const;
+    int32_t GetSize(int64_t &size) const;
     int32_t CallbackCheckAndSetNoChange();
     void Release();
 
@@ -49,7 +49,7 @@ private:
     std::shared_ptr<CallbackWorks> callbackWorks_ = nullptr;
     std::shared_ptr<JsCallback> readAt_ = nullptr;
     std::shared_ptr<JsCallback> getMem_ = nullptr;
-    int32_t size_ = -1;
+    int64_t size_ = -1;
     bool noChange_ = false;
 };
 } // namespace Media
